+++ ArtifactOfPrestige/ArtifactBase.cs
@@ -5,6 +5,7 @@ using System;
 using System.Collections.Generic;
 using System.Text;
 using UnityEngine;
+using UnityEngine.Networking;
 
 namespace ArtifactOfPrestige
 {